backend/requirements.txt
flask
flask-cors
google-genai
edge-tts
pydub
